Refactor opacity

+++ migrations/2025-01-24-140656_user_settings/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE user_settings (
     username VARCHAR NOT NULL,
-    background_opacity INTEGER NOT NULL,
+    background_opacity REAL NOT NULL,
     fps_target INTEGER NOT NULL,
     PRIMARY KEY(username),
     FOREIGN KEY(username) REFERENCES users(username)
